Added wrapper constants on castTo
The castTo operations is used on field declarion processing, and when a convertion between wrappers and primitives, the constant value was lost. The situation that shows the problem is explained on https://www.eclipse.org/forums/index.php/t/1104636/.

This change makes possible the casting between primitives and wrappers returning "this" as well as done with casting with same target and from types.

The only exception here is "byte", that have a colision with other switch flag.

Signed-off-by: Marcelo Avancini <marcelo.avan@gmail.com>